refactor: Update axios and api classes to use abstracted instance and add support for user avatar
This commit introduces changes to the API and Section classes to abstract the `axios` instance and use it consistently for every method in the class. It also adds support for fetching user avatars using the new `/users.getAvatar` API endpoint. The `Profile` and `User` components have been updated to use the new `users.getAvatar` method from the `api` class to fetch user avatars.